pyproject.toml
[project]
name = "megacontext"
version = "0.1.0"
description = "MegaContext proof-of-concept runtime"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [{name = "MegaContext Contributors"}]
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.39.0",
    "accelerate>=0.27.0",
    "datasets>=2.15.0",
    "wandb>=0.16.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "pyarrow>=14.0.2",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.1.0",
    "ruff>=0.3.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.6.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]
skip-string-normalization = true

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src"]
exclude = ["build"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
branch = true
source = ["megacontext"]
